requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
celery==5.5.2
django-celery-beat==2.8.0
redis==5.2.1